+++ NYear.ODA/Adapters/DbASybase.cs
@@ -123,7 +123,7 @@ namespace NYear.ODA.Adapter
         {
             return null;
         }
-        public override DatabaseColumnInfo ODAColumnToOrigin(string Name, string ColumnType, decimal Length)
+        public override DatabaseColumnInfo ODAColumnToOrigin(string Name, string ColumnType, int Length)
         {
             DatabaseColumnInfo ColInof = new DatabaseColumnInfo();
             ColInof.Name = Name;
